requirements.txt
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.3.2
coverage==7.3.2
docutils==0.20.1
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
packaging==23.2
parse==1.19.1
pkginfo==1.9.6
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
pytest==7.4.3
pytest-cov==4.1.0
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
requests-wsgi-adapter==0.4.1
rfc3986==2.0.0
rich==13.6.0
twine==4.0.2
urllib3==2.0.7
WebOb==1.8.7
whitenoise==6.6.0
zipp==3.17.0
